+++ CHANGE_LOG.md
@@ -31,3 +31,12 @@ Implemented
 Implemented
     Clickable email and phone in contact-modal(footer.ejs & landing-nav.ejs)
     - Deepikaa
+Implemented
+    Pooja booking controller jest test
+    - Miruthul
+Implemented
+    Payment Controller Jest Test
+    - Miruthul
+Implemented
+    Family Controller Jest Test
+    - Miruthul